+++ database/migrations/2021_10_08_143103_create_laporan_produksi_detail_table.php
@@ -21,7 +21,7 @@ class CreateLaporanProduksiDetailTable extends Migration
             $table->string('nomor_laporan');
             $table->decimal('berat', 5, 3);
             $table->integer('qty');
-            $table->integer('total');
+            $table->float('total', 7, 3);
             $table->timestamps();
         });
     }
